+++ src/Servers/Logger.php
@@ -7,6 +7,7 @@ use Innmind\Server\Control\{
     Server,
     Server\Processes,
     Server\Volumes,
+    Server\Command,
 };
 use Innmind\Immutable\Attempt;
 use Psr\Log\LoggerInterface;
@@ -47,12 +48,12 @@ final class Logger implements Server
     #[\Override]
     public function reboot(): Attempt
     {
-        return Server\Script::of('sudo shutdown -r now')($this);
+        return Server\Script::of(Command::foreground('sudo shutdown -r now'))($this);
     }
 
     #[\Override]
     public function shutdown(): Attempt
     {
-        return Server\Script::of('sudo shutdown -h now')($this);
+        return Server\Script::of(Command::foreground('sudo shutdown -h now'))($this);
     }
 }