+++ CHANGELOG.md
@@ -33,6 +33,7 @@
     - `Innmind\Server\Control\Server\Command::outputToBeStreamed()`
     - `Innmind\Server\Control\Server\Command::toString()`
     - `Innmind\Server\Control\Servers\Unix::of()`
+- `Innmind\Server\Control\Server\Script` constructor is now private, use `::of()` named constructor which now expects instances of `Command`
 
 ### Removed